DTSPB-2626 Fix s2s_auth & acc_manageCase Pact Test
